+++ spec/lib/filter/query_spec.rb
@@ -289,7 +289,7 @@ describe Filter::Query do
         ).query_full
       }.to raise_error(CustomErrors::FilterArgumentError, /Projections hash must have exactly 1 entry, got 2/)
     end
- 
+
     it 'occurs when projection has empty include' do
       expect {
         create_filter(
@@ -307,6 +307,14 @@ describe Filter::Query do
       }.to raise_error(CustomErrors::FilterArgumentError, /Include must contain at least one field/)
     end
 
+    it 'occurs with a deformed \'in\' filter' do
+      filter_params = {'filter' => {'siteId' => {'in' => [{'customLatitude' => nil,'customLongitude' => nil,'description' => nil,'id' => 508,'locationObfuscated' => true,'name' => 'Site 1','projectIds' => [397],'links' => ['http://example.com/projects/397/sites/508']},{'customLatitude' => nil,'customLongitude' => nil,'description' => nil,'id' => 400,'locationObfuscated' => true,'name' => 'Site 2','projectIds' => [397],'links' => ['http://example.com/projects/397/sites/400']},{'customLatitude' => nil,'customLongitude' => nil,'description' => nil,'id' => 402,'locationObfuscated' => true,'name' => 'Site 3','projectIds' => [397],'links' => ['http://example.com/projects/397/sites/402']},{'customLatitude' => nil,'customLongitude' => nil,'description' => nil,'id' => 399,'locationObfuscated' => true,'name' => 'Site 4','projectIds' => [397,469],'links' => ['http://example.com/projects/397/sites/399','http://example.com/projects/469/sites/399']},{'customLatitude' => nil,'customLongitude' => nil,'description' => nil,'id' => 401,'locationObfuscated' => true,'name' => 'Site 5','projectIds' => [397],'links' => ['http://example.com/projects/397/sites/401']},{'customLatitude' => nil,'customLongitude' => nil,'description' => nil,'id' => 398,'locationObfuscated' => true,'name' => 'Site 6','projectIds' => [397,469],'links' => ['http://example.com/projects/397/sites/398','http://example.com/projects/469/sites/398']}]}},'projection' => {'include' => ['id','siteId','durationSeconds','recordedDate']}}
+      
+      expect {
+        create_filter(filter_params).query_full
+      }.to raise_error(CustomErrors::FilterArgumentError, 'Array values cannot be hashes.')
+    end
+    
   end
 
   context 'projection' do