+++ CHANGELOG.md
@@ -2,6 +2,16 @@
 
 ## Not Released
 
+ - 2015-01-18
+    - delete actions improved, archive headers added and fixed
+    - standardised controller authorisation
+    - added audio event filter action
+
+ - 2015-01-06
+    - Fixed CORS responses.
+    - Added ability to poll Resque for job completion rather than polling filesystem.
+    - Bug fix: added more strict validation and more tests for 'in' filter.
+
  - 2014-12-30
     - Upgraded to Rails 4.2.0.
     - Migrated from protected attributes to strong parameters.